Fix ratings records

+++ octane/pipelines/game_records.go
@@ -36,7 +36,7 @@ func GamePlayerRecords(filter bson.M, stat string) *Pipeline {
 				Opponent *octane.Team   `json:"opponent,omitempty" bson:"opponent,omitempty"`
 				Winner   bool           `json:"winner,omitempty" bson:"winner,omitempty"`
 				Player   *octane.Player `json:"player,omitempty" bson:"player,omitempty"`
-				Stat     int            `json:"stat,omitempty" bson:"stat,omitempty"`
+				Stat     float64        `json:"stat,omitempty" bson:"stat,omitempty"`
 			}
 			if err := cursor.Decode(&player); err != nil {
 				return nil, err
